Refactored CargoManager
