+++ src/components/cargo/BuildType.ts
@@ -1,0 +1,4 @@
+export enum BuildType {
+    Debug,
+    Release
+}
