+++ src/components/cargo/CheckTarget.ts
@@ -1,0 +1,4 @@
+export enum CheckTarget {
+    Library,
+    Application
+}
