+++ src/components/cargo/CrateType.ts
@@ -1,0 +1,4 @@
+export enum CrateType {
+    Application,
+    Library
+}
